// seall-web/src/api/generated/endpoints.ts
export const API_ENDPOINTS = {
// ...
    MEDIA_LIBRARY: {
        /**
         *  @description
         *  Route returns the main local anime collection.
         *  This creates a new LibraryCollection struct and returns it.
         *  This is used to get the main anime collection of the user.
         *  It uses the cached Simkl anime collection for the GET method.
         *  It refreshes the SIMKL anime collection if the POST method is used.
         */
        GetLibraryCollection: {
            key: "MEDIA-LIBRARY-get-library-collection",
            methods: ["GET", "POST"],
            endpoint: "/api/v1/library/collection",
        },
        /**
         *  @description
         *  Route returns anime collection schedule
         *  This is used by the "Schedule" page to display the anime schedule.
         */
        GetMediaCollectionSchedule: {
            key: "MEDIA-LIBRARY-get-media-collection-schedule",
            methods: ["GET"],
            endpoint: "/api/v1/library/schedule",
        },
        /**
         *  @description
         *  Route adds the given media to the user's SIMKL planning collections
         *  Since media not found in the user's SIMKL collection are not displayed in the library, this route is used to add them.
         *  The response is ignored in the frontend, the client should just refetch the entire library collection.
         */
        AddUnknownMedia: {
            key: "MEDIA-LIBRARY-add-unknown-media",
            methods: ["POST"],
            endpoint: "/api/v1/library/unknown-media",
        },
    },
// ...
} satisfies ApiEndpoints